// EasyPayFrontEnd/src/app/model/WalletAccount.ts
/**
 * Class which define the entites of WalletAccount table.
 */
export class WalletAccount {
    firstName: string;
    middleName: string;
    lastName: string;
    gender: string;
    mobile_no: string;
    emailId: string;
    user_name: string;
    password: string;
    dob: string;
    balance: number;
}